This is a template file for making commits to the ckbs repository. Lines with no 'at' characters, are general comments not connected to a specifi file. Lines containing an 'at' character are "file name" followed by comment. Next line must be empty for ./commit.sh files to work.
commit.sh: For this example, commit.sh is the only file committed.

git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@1738 0e562018-b8fb-0310-a933-ecf073c7c197

+++ commit.sh
@@ -20,13 +20,16 @@ then
 	list=`cat list.$$`
 	for file in $list
 	do
-		sed -f svn_commit.sed $file > commit.$$
-		if ! diff $file commit.$$ > /dev/null
+		if [ -e "$file" ]
 		then
-			echo "abort commit.sh: suggest following changes to $file:"
-			diff $file commit.$$
-			rm commit.$$
-			exit 1
+			sed -f svn_commit.sed $file > commit.$$
+			if ! diff $file commit.$$ > /dev/null
+			then
+				echo "commit.sh aborting: suggest changes to $file:"
+				diff $file commit.$$
+				rm commit.$$
+				exit 1
+			fi
 		fi
 	done
 	#
